Split `ArrayShardedExt` and `ArrayShardedReadableExt`
