+++ src/array.rs
@@ -27,6 +27,8 @@ mod unsafe_cell_slice;
 
 #[cfg(feature = "sharding")]
 mod array_sharded_ext;
+#[cfg(feature = "sharding")]
+mod array_sync_sharded_readable_ext;
 
 use std::sync::Arc;
 
@@ -53,9 +55,9 @@ pub use self::{
 };
 
 #[cfg(feature = "sharding")]
-pub use array_sharded_ext::{
-    ArrayShardedExt, ArrayShardedReadableExt, ArrayShardedReadableExtCache,
-};
+pub use array_sharded_ext::ArrayShardedExt;
+#[cfg(feature = "sharding")]
+pub use array_sync_sharded_readable_ext::{ArrayShardedReadableExt, ArrayShardedReadableExtCache};
 
 use serde::Serialize;
 use thiserror::Error;
